@@ -67,7 +67,7 @@ var _ = ginkgo.Describe("[Feature: Real IP]", func() {
 					rd := bufio.NewReader(c)
 					ppHeader, err := pp.Read(rd)
 					if err != nil {
-						log.Error("read proxy protocol error: %v", err)
+						log.Errorf("read proxy protocol error: %v", err)
 						return
 					}
 
@@ -95,7 +95,7 @@ var _ = ginkgo.Describe("[Feature: Real IP]", func() {
 			f.RunProcesses([]string{serverConf}, []string{clientConf})
 
 			framework.NewRequestExpect(f).Port(remotePort).Ensure(func(resp *request.Response) bool {
-				log.Trace("ProxyProtocol get SourceAddr: %s", string(resp.Content))
+				log.Tracef("ProxyProtocol get SourceAddr: %s", string(resp.Content))
 				addr, err := net.ResolveTCPAddr("tcp", string(resp.Content))
 				if err != nil {
 					return false
@@ -123,7 +123,7 @@ var _ = ginkgo.Describe("[Feature: Real IP]", func() {
 					rd := bufio.NewReader(c)
 					ppHeader, err := pp.Read(rd)
 					if err != nil {
-						log.Error("read proxy protocol error: %v", err)
+						log.Errorf("read proxy protocol error: %v", err)
 						return
 					}
 					srcAddrRecord = ppHeader.SourceAddr.String()
@@ -145,7 +145,7 @@ var _ = ginkgo.Describe("[Feature: Real IP]", func() {
 				r.HTTP().HTTPHost("normal.example.com")
 			}).Ensure(framework.ExpectResponseCode(404))
 
-			log.Trace("ProxyProtocol get SourceAddr: %s", srcAddrRecord)
+			log.Tracef("ProxyProtocol get SourceAddr: %s", srcAddrRecord)
 			addr, err := net.ResolveTCPAddr("tcp", srcAddrRecord)
 			framework.ExpectNoError(err, srcAddrRecord)
 			framework.ExpectEqualValues("127.0.0.1", addr.IP.String())
